+++ tests/testthat/test-bas.lm.R
@@ -83,28 +83,6 @@ test_that("prediction versus fitted", {
                as.vector(predict(hald.ZS, estimator="MPM")$fit))
 })
 
-test_that("GLM logit probne0", {
-  data(Pima.tr, package="MASS")
-  pima.BAS = bas.glm(type ~ ., data=Pima.tr, n.models= 2^7,
-                     method="BAS",
-                     betaprior=bic.prior(), family=binomial(),
-                     modelprior=uniform())
-  pima.det = bas.glm(type ~ ., data=Pima.tr, n.models= 2^7,
-                     method="deterministic",
-                     betaprior=bic.prior(), family=binomial(),
-                     modelprior=uniform())
-  expect_equal(pima.BAS$probne0, pima.BAS$probne0)
-  pima.BAS = bas.glm(type ~ ., data=Pima.tr, n.models= 2^7,
-                     method="BAS",
-                     betaprior=Jeffreys(), family=binomial(),
-                     modelprior=uniform())
-  pima.det = bas.glm(type ~ ., data=Pima.tr, n.models= 2^7,
-                     method="deterministic",
-                     betaprior=Jeffreys(), family=binomial(),
-                     modelprior=uniform())
-  expect_equal(pima.BAS$probne0, pima.BAS$probne0)
-
-})
 
 test_that("force.heredity", {
   loc = system.file("testdata", package="BAS")